requirements.txt
torch>=2.0
transformers>=4.39
numpy>=1.24
pyyaml>=6.0
tqdm>=4.66 